requirements.txt
streamlit>=1.22.0
langchain>=0.3.7
langchain-community>=0.3.7
langchain-google-genai>=2.0.5
chromadb>=0.5.11
python-dotenv>=1.0.1
pypdf>=5.0.0
